backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
sqlalchemy==2.0.23
pydantic==2.8.2
python-dotenv==1.0.0
httpx==0.25.2
python-multipart==0.0.6
python-docx==1.1.0
PyPDF2==3.0.1
requests==2.31.0

# SDK
openai==1.3.0
anthropic==0.7.0
deepseek-sdk==0.1.1